+++ R/zzz.R
@@ -6,7 +6,8 @@
     crssio.traceNumberSlot = 'HydrologyParameters.TraceNumber',
     crssio.hydroIncrement = 'MeadFloodControlData.hydrologyIncrement',
     crssio.sacYTSlot = "MWD_ICS.SacWYType",
-    crssio.histNfFile = "HistoricalNaturalFlow.xlsx"
+    crssio.histNfFile = "HistoricalNaturalFlow.xlsx",
+    crssio.nf_sheet_name = "InterveningNaturalFlow"
   )
   toset <- !(names(op.crssio) %in% names(op))
   if(any(toset)) options(op.crssio[toset])